Fix old code to use traditional hooks with Axios
I will leave the useAxios code commented.
The problem with my old code was that I'd setGifs to the specific URL
instead of response.data (the object array), and that's why I was
struggling to access the URLs in the return. I also fixed the
dependencies to be set to [setGifs] rather than [gifs] to prevent
infinite rendering.